QuestionsAndAnswers
Elementary

Write a program that prints ‘Hello World’ to the screen.
Write a program that asks the user for her name and greets her with her name.
Modify the previous program such that only the users Alice and Bob are greeted with their names.
Write a program that asks the user for a number n and prints the sum of the numbers 1 to n
Modify the previous program such that only multiples of three or five are considered in the sum, e.g. 3, 5, 6, 9, 10, 12, 15 for n=17
Write a program that asks the user for a number n and gives him the possibility to choose between computing the sum and computing the product of 1,…,n.
Write a program that prints a multiplication table for numbers up to 12.
Write a program that prints all prime numbers. (Note: if your programming language does not support arbitrary size numbers, printing all primes up to the largest number you can easily represent is fine too.)
Write a guessing game where the user has to guess a secret number. After every guess the program tells the user whether their number was too large or too small. At the end the number of tries needed should be printed. I counts only as one try if they input the same number multiple times consecutively.
Write a program that prints the next 20 leap years.
Write a program that computes 4⋅∑k=1106(−1)k+12k−1=4⋅(1−1/3+1/5−1/7+1/9−1/11…).4\cdot \sum_{k=1}^{10^6} \frac{(-1)^{k+1}}{2k-1} = 4\cdot(1-1/3+1/5-1/7+1/9-1/11\ldots).
Lists, Strings

If your language of choice doesn’t have a build in list and/or string type (e.g. you use C), these exercises should also be solvable for arrays. However, some solutions are very different between an array-based list (like C++’s vector) and a pointer based list (like C++’s list), at least if you care about the efficiency of your code. So you might want to either find a library, or investigate how to implement your own linked list if your language doesn’t have it.
 
Write a function that returns the largest element in a list.
Write function that reverses a list, preferably in place.
Write a function that checks whether an element occurs in a list.
Write a function that returns the elements on odd positions in a list.
Write a function that computes the running total of a list.
Write a function that tests whether a string is a palindrome.
Write three functions that compute the sum of the numbers in a list: using a for-loop, a while-loop and recursion. (Subject to availability of these constructs in your language of choice.)
Write a function on_all that applies a function to every element of a list. Use it to print the first twenty perfect squares (a natural number nn is a perfect square if it can be written as n=m*mn=m*m for some other natural number mm. 1,4,9,16,251,4,9,16,25 are the first 5).
Write a function that concatenates two lists. [a,b,c],[1,2,3] → [a,b,c,1,2,3]
Write a function that combines two lists by alternatingly taking elements, e.g. [a,b,c], [1,2,3] →[a,1,b,2,c,3].
Write a function that merges two sorted lists into a new sorted list. [1,4,6],[2,3,5] → [1,2,3,4,5,6]. You can do this quicker than concatenating them followed by a sort.
Write a function that rotates a list by k elements. For example [1,2,3,4,5,6] rotated by two becomes[3,4,5,6,1,2]. Try solving this without creating a copy of the list. How many swap or move operations do you need?
Write a function that computes the list of the first 100 Fibonacci numbers.
Write a function that takes a number and returns a list of its digits.
Write functions that add, subtract, and multiply two numbers in their digit-list representation (and return a new digit list). If you’re ambitious you can implement Karatsuba multiplication. Try different bases. What is the best base if you care about speed? If you couldn’t completely solve the prime number exercise above due to the lack of large numbers in your language, you can now use your own library for this task.
Write a function that takes a list of numbers, a starting base b1 and a target base b2 and interprets the list as a number in base b1 and converts it into a number in baseb2 (in the form of a list-of-digits).
Implement the following sorting algorithms: Selection sort, Insertion sort, Merge sort, Quick sort, Stooge Sort. Check Wikipedia for descriptions.
Implement binary search.
19.  Write a function that takes a list of strings an prints them, one per line, in a rectangular frame. For example the list ["Hello", "World", "in", "a", "frame"] gets printed as:
20. *********
21. * Hello *
22. * World *
23. * in    *
24. * a     *
25. * frame *
*********
26.  Write function that translates a text to Pig Latin and back. English is translated to Pig Latin by taking the first letter of every word, moving it to the end of the word and adding ‘ay’. “The quick brown fox” becomes “Hetay uickqay rownbay oxfay”.